+++ client/src/components/molecules/DatePicker/DatePicker.stories.tsx
@@ -1,0 +1,11 @@
+import React from 'react';
+import DatePicker from './DatePicker';
+
+export default {
+  title: 'Molecules/DatePicker',
+  component: DatePicker,
+};
+
+export const Default = () => {
+  return <DatePicker />;
+};
